+++ src/main/java/act/job/JobManager.java
@@ -20,6 +20,8 @@ package act.job;
  * #L%
  */
 
+import static act.app.event.SysEventId.APP_CODE_SCANNED;
+
 import act.Act;
 import act.Destroyable;
 import act.app.*;
@@ -32,6 +34,7 @@ import act.util.SimpleProgressGauge;
 import org.joda.time.DateTime;
 import org.joda.time.Seconds;
 import org.osgl.$;
+import org.osgl.cache.CacheService;
 import org.osgl.exception.NotAppliedException;
 import org.osgl.logging.LogManager;
 import org.osgl.logging.Logger;
@@ -53,17 +56,24 @@ public class JobManager extends AppServiceBase<JobManager> {
     private ConcurrentMap<String, Job> jobs = new ConcurrentHashMap<String, Job>();
     private ConcurrentMap<Method, Job> methodIndex = new ConcurrentHashMap<>();
     private ConcurrentMap<String, ScheduledFuture> scheduled = new ConcurrentHashMap<>();
+    private CacheService jobResultCache;
 
     static String sysEventJobId(SysEventId eventId) {
         return S.concat(SYS_JOB_MARKER, eventId.toString().toLowerCase());
     }
 
-    public JobManager(App app) {
+    public JobManager(final App app) {
         super(app);
         initExecutor(app);
         for (SysEventId sysEventId : SysEventId.values()) {
             createSysEventListener(sysEventId);
         }
+        on(APP_CODE_SCANNED, "init-job-result-cache", new Runnable() {
+            @Override
+            public void run() {
+                jobResultCache = app.cache("job_result_cache");
+            }
+        });
     }
 
     @Override
@@ -136,6 +146,10 @@ public class JobManager extends AppServiceBase<JobManager> {
     public String prepare($.Function<ProgressGauge, ?> worker) {
         Job job = wrap(worker);
         return job.id();
+    }
+
+    public void prepare(String jobId, $.Function<ProgressGauge, ?> worker) {
+        wrap(jobId, worker);
     }
 
     /**
@@ -417,6 +431,14 @@ public class JobManager extends AppServiceBase<JobManager> {
         if (null != method) {
             methodIndex.remove(method);
         }
+    }
+
+    public void cacheResult(String jobId, Object result) {
+        jobResultCache.put("__jr_" + jobId, result, 60);
+    }
+
+    public Object cachedResult(String jobId) {
+        return jobResultCache.get("__jr_" + jobId);
     }
 
     ScheduledThreadPoolExecutor executor() {